brain-zero/requirements.txt
# brain-zero: from-scratch LLM training (Python 3.10+)
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
tokenizers>=0.15.0
accelerate>=0.26.0
pyyaml>=6.0.1
tqdm>=4.66.0
numpy>=1.24.0
